Init: Create odin-recipes index.html and README.md
Creating a directory for T.O.P. projects & the Recipes project.

Created a homepage for the recipes project using a VSCode boilerplate,
as well as a basic README to act as a fill-in until I flesh it out at
the end of the project.

Curious to see how multiline commit messages like this will look like. Never
actually done it before.